% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melsi_robust.R
\name{plot_pcoa}
\alias{plot_pcoa}
\title{Plot PCoA from MeLSI Results}
\usage{
plot_pcoa(melsi_results, X, y, title = "PCoA using MeLSI Distance")
}
\arguments{
\item{melsi_results}{Results object from melsi() function}

\item{X}{Original feature matrix (samples x taxa)}

\item{y}{Group labels vector}

\item{title}{Optional custom title for the plot (default: "PCoA using MeLSI Distance")}
}
\value{
A ggplot2 object (invisibly)
}
\description{
Creates a Principal Coordinates Analysis (PCoA) plot using the learned MeLSI distance matrix.
}
\examples{
\dontrun{
  # Run MeLSI analysis
  results <- melsi(X_clr, y)
  
  # Plot PCoA (automatic!)
  plot_pcoa(results, X_clr, y)
  
  # Custom title
  plot_pcoa(results, X_clr, y, title = "My PCoA Plot")
}

}
